Create GraphicOutput.java
Initial Commit of the main class for Optera Iconic 3.0. This is the class that will initialize the main drawing window and also hold the methods for drawing various items within it. Unlike previous versions the intent of this version is to be as efficient as possible with processor cycles while still retaining lots of functionality.